migrate articleBodyAdverts away from commercialFeatures
