+++ bundle/src/insert/spacefinder/article.spec.ts
@@ -1,4 +1,4 @@
-import { commercialFeatures } from '../../lib/commercial-features';
+import { allowArticleBodyAdverts } from '../../lib/article-body-adverts';
 import { init } from './article';
 import { spaceFiller } from './space-filler';
 
@@ -10,8 +10,8 @@ jest.mock('insert/fill-dynamic-advert-slot', () => ({
 	fillDynamicAdSlot: jest.fn(),
 }));
 
-jest.mock('lib/commercial-features', () => ({
-	commercialFeatures: {},
+jest.mock('lib/article-body-adverts', () => ({
+	allowArticleBodyAdverts: jest.fn().mockReturnValue(true),
 }));
 
 jest.mock('insert/spacefinder/space-filler', () => ({
@@ -38,8 +38,6 @@ const mockViewport = (width: number, height: number): void => {
 describe('Article Body Adverts', () => {
 	beforeEach(() => {
 		jest.resetAllMocks();
-		commercialFeatures.articleBodyAdverts = true;
-		// @ts-expect-error -- we need to TS space-filler’s queue
 		spaceFillerStub.mockImplementation(() => Promise.resolve(2));
 		mockViewport(0, 1300);
 		expect.hasAssertions();
@@ -49,9 +47,9 @@ describe('Article Body Adverts', () => {
 		expect(init).toBeDefined();
 	});
 
-	it('should exit if commercial feature disabled', () => {
+	it('should exit if ads in article body are disabled', () => {
 		const fillAdSlot = jest.fn();
-		commercialFeatures.articleBodyAdverts = false;
+		jest.mocked(allowArticleBodyAdverts).mockReturnValue(false);
 		return init(fillAdSlot).then(() => {
 			expect(spaceFillerStub).not.toHaveBeenCalled();
 		});
@@ -59,6 +57,7 @@ describe('Article Body Adverts', () => {
 
 	it('should call relevant functions to fill space on desktop', () => {
 		const fillAdSlot = jest.fn();
+		jest.mocked(allowArticleBodyAdverts).mockReturnValue(true);
 		mockViewport(1300, 1300);
 		return init(fillAdSlot).then(() => {
 			expect(spaceFillerStub).toHaveBeenCalledTimes(2);
@@ -72,6 +71,7 @@ describe('Article Body Adverts', () => {
 
 	it('should call relevant functions to fill space on mobile and tablet', () => {
 		const fillAdSlot = jest.fn();
+		jest.mocked(allowArticleBodyAdverts).mockReturnValue(true);
 		mockViewport(500, 1300);
 		return init(fillAdSlot).then(() => {
 			expect(spaceFillerStub).toHaveBeenCalledTimes(1);
